putting tracking component in separate file
